(PC-27682)[PRO] refactor: merge validation schemas
