--- pro/src/components/VenueForm/Accessibility/index.ts
@@ -1,2 +1,0 @@
-export { default as validationSchema } from './validationSchema'
-export { default as Accessibility } from './Accessibility'
