--- pro/src/components/VenueForm/Activity/index.ts
@@ -1,1 +1,0 @@
-export { default as validationSchema } from './validationSchema'
