--- pro/src/components/VenueForm/Activity/validationSchema.ts
@@ -1,9 +1,0 @@
-import * as yup from 'yup'
-
-const validationSchema = {
-  venueType: yup
-    .string()
-    .required('Veuillez sélectionner une activité principale'),
-}
-
-export default validationSchema
