--- pro/src/components/VenueForm/Contact/index.ts
@@ -1,1 +1,0 @@
-export { default as validationSchema } from './validationSchema'
