+++ pro/src/components/VenueForm/index.ts
@@ -1,3 +1,2 @@
-export { default as validationSchema } from './validationSchema'
 export type { VenueFormValues } from './types'
 export { setDefaultInitialFormValues, setInitialFormValues } from './utils'
