+++ pro/src/screens/VenueForm/VenueCreationFormScreen.tsx
@@ -5,8 +5,9 @@ import { useLocation, useNavigate } from 'react-router-dom'
 import { api } from 'apiClient/api'
 import { isErrorAPIError, serializeApiErrors } from 'apiClient/helpers'
 import { GetOffererResponseModel, VenueProviderResponse } from 'apiClient/v1'
-import { validationSchema, VenueFormValues } from 'components/VenueForm'
+import { VenueFormValues } from 'components/VenueForm'
 import { generateSiretValidationSchema } from 'components/VenueForm/Informations/SiretOrCommentFields'
+import { validationSchema } from 'components/VenueForm/validationSchema'
 import { VenueCreationForm } from 'components/VenueForm/VenueCreationForm'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
