+++ pro/src/screens/VenueForm/VenueEditionFormScreen.tsx
@@ -7,8 +7,9 @@ import { isErrorAPIError, serializeApiErrors } from 'apiClient/helpers'
 import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import Callout from 'components/Callout/Callout'
 import ConfirmDialog from 'components/Dialog/ConfirmDialog'
-import { validationSchema, VenueFormValues } from 'components/VenueForm'
+import { VenueFormValues } from 'components/VenueForm'
 import { generateSiretValidationSchema } from 'components/VenueForm/Informations/SiretOrCommentFields'
+import { validationSchema } from 'components/VenueForm/validationSchema'
 import { VenueEditionForm } from 'components/VenueForm/VenueEditionForm'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
